Created convenience on CATransaction for chaining
A nested `Animation` class holds the parameters for a particular
animation. In particular, it has an enum, `Completion`, that describes
what to do when the animation is finished. This can be either a simple
function to run, or another `Animation` instance.

`CATransaction` gains a static `animate(_:Animation)` function to kick
off the process, and unwrap and invoke any chained `Animation`s.

Added a README.